Simplify net.Listener/tls.Listener handling.
