+++ rproxy/rproxy.go
@@ -175,3 +175,31 @@ func (proxy *Proxy) enterExecution() {
 func (proxy *Proxy) leaveExecution() {
 	proxy.channels.releasePermission <- struct{}{}
 }
+
+func (proxy *Proxy) startListening() error {
+	ln, addr, err := proxy.config.Listen.Listen()
+	if err != nil {
+		return err
+	}
+	proxy.listenAddr = addr
+	go proxy.listenForClients(ln)
+	return nil
+}
+
+func (proxy *Proxy) listenForClients(ln *Listener) {
+	defer ln.Close()
+
+	for proxy.State().IsStartingOrAlive() {
+		ln.SetDeadline(time.Now().Add(time.Second))
+		conn, err := ln.Accept()
+		if err != nil {
+			if resp.IsNetTimeout(err) {
+				continue
+			}
+			log.Printf("Got an error accepting a connection: %s", err)
+		} else {
+			rc := resp.NewConn(conn, 0, proxy.config.LogMessages)
+			go NewCliHandler(rc, proxy).Run()
+		}
+	}
+}
